test(interaction): split Telegram authorization suite into its own file
telegram.test.ts sat at 797 of the 800-line test limit, leaving no room for
the drain-termination and chat-id normalization regression tests. Pure move --
no test body changes. Naming follows the split-by-describe-block convention in
.claude/rules/test-architecture.md.

Both suites now reach the fetch seam through the @/interaction barrel instead
of a 4-level relative path, so _telegramPluginDeps is exported from the barrel
alongside the plugin. That retires the pre-existing deep relative as well --
deep-relative baseline 2847 -> 2846.
