+++ scripts/baselines/deep-relatives-baseline.json
@@ -1,6 +1,6 @@
 {
-  "count": 2847,
-  "updatedAt": "2026-07-18T08:28:39.440Z",
+  "count": 2846,
+  "updatedAt": "2026-07-28T05:07:08.700Z",
   "byFile": {
     "src/pipeline/stages/acceptance-setup.ts": 1,
     "src/pipeline/stages/prompt.ts": 5,
@@ -644,7 +644,7 @@
     "test/unit/interaction/triggers-narrowed.test.ts": 4,
     "test/unit/interaction/plugins/cli.test.ts": 1,
     "test/unit/interaction/chain.test.ts": 2,
-    "test/unit/interaction/interaction-plugins.test.ts": 6,
+    "test/unit/interaction/interaction-plugins.test.ts": 5,
     "test/unit/interaction/triggers.test.ts": 4,
     "test/unit/interaction/telegram-timeout.test.ts": 2,
     "test/unit/interaction/auto-plugin-adapter.test.ts": 3,
@@ -765,13 +765,13 @@
     "test/helpers/mock-session-manager.ts": 2,
     "test/helpers/mock-story.ts": 1,
     "test/helpers/review-audit.ts": 2,
+    "test/helpers/warn-spy.ts": 1,
     "test/helpers/pipeline-context.ts": 4,
     "test/helpers/fake-agent-manager.ts": 10,
     "test/helpers/mock-nax-config.ts": 2,
     "test/helpers/helpers.test.ts": 2,
     "test/helpers/index.ts": 1,
     "test/helpers/deps.ts": 1,
-    "test/helpers/verbatim-warn-spy.ts": 1,
     "test/helpers/mock-agent-adapter.ts": 1,
     "scripts/logging-formatter-demo.ts": 3
   }
